+++ codeJK2/cgame/cg_scoreboard.cpp
@@ -240,7 +240,7 @@ w = cgi_R_Font_StrLenPixels(text, cgs.media.qhFontSmall, 0.8f);
 	const float percent = cg_entities[0].gent->client->sess.missionStats.shotsFired? 100.0f * (float)cg_entities[0].gent->client->sess.missionStats.hits / cg_entities[0].gent->client->sess.missionStats.shotsFired : 0;
 	cgi_R_Font_DrawString(x+w, y, va("%.2f%%",percent), colorTable[CT_WHITE], cgs.media.qhFontSmall, -1, 0.8f);
 
-	if ( cg_entities[0].gent->client->sess.missionStats.weaponUsed[WP_SABER] <= 0 )
+	if ( cg_entities[0].gent->client->sess.missionStats.weaponUsed[static_cast<int>(WP_SABER)] <= 0 )
 	{
 		return; //don't have saber yet, so don't print any stats
 	}
