+++ codeJK2/game/wp_bowcaster.cpp
@@ -35,7 +35,7 @@ along with this program; if not, see <http://www.gnu.org/licenses/>.
 static void WP_BowcasterMainFire( gentity_t *ent )
 //---------------------------------------------------------
 {
-	int			damage	= weaponData[WP_BOWCASTER].damage, count;
+	int			damage	= weaponData[static_cast<int>(WP_BOWCASTER)].damage, count;
 	float		vel;
 	vec3_t		angs, dir, start;
 	gentity_t	*missile;
@@ -118,8 +118,8 @@ static void WP_BowcasterMainFire( gentity_t *ent )
 		missile->dflags = DAMAGE_DEATH_KNOCKBACK;
 		missile->methodOfDeath = MOD_BOWCASTER;
 		missile->clipmask = MASK_SHOT | CONTENTS_LIGHTSABER;
-		missile->splashDamage = weaponData[WP_BOWCASTER].splashDamage;
-		missile->splashRadius = weaponData[WP_BOWCASTER].splashRadius;
+		missile->splashDamage = weaponData[static_cast<int>(WP_BOWCASTER)].splashDamage;
+		missile->splashRadius = weaponData[static_cast<int>(WP_BOWCASTER)].splashRadius;
 
 		// we don't want it to bounce
 		missile->bounceCount = 0;
@@ -132,7 +132,7 @@ static void WP_BowcasterAltFire( gentity_t *ent )
 //---------------------------------------------------------
 {
 	vec3_t	start;
-	int		damage	= weaponData[WP_BOWCASTER].altDamage;
+	int		damage	= weaponData[static_cast<int>(WP_BOWCASTER)].altDamage;
 
 	VectorCopy( wpMuzzle, start );
 	WP_TraceSetStart( ent, start, vec3_origin, vec3_origin );//make sure our start point isn't on the other side of a wall
@@ -176,8 +176,8 @@ static void WP_BowcasterAltFire( gentity_t *ent )
 	missile->dflags = DAMAGE_DEATH_KNOCKBACK;
 	missile->methodOfDeath = MOD_BOWCASTER_ALT;
 	missile->clipmask = MASK_SHOT | CONTENTS_LIGHTSABER;
-	missile->splashDamage = weaponData[WP_BOWCASTER].splashDamage;
-	missile->splashRadius = weaponData[WP_BOWCASTER].splashRadius;
+	missile->splashDamage = weaponData[static_cast<int>(WP_BOWCASTER)].splashDamage;
+	missile->splashRadius = weaponData[static_cast<int>(WP_BOWCASTER)].splashRadius;
 }
 
 //---------------------------------------------------------
